requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.26
langchain-google-genai==2.0.8
langgraph==0.2.70
google-generativeai==0.8.3
faiss-cpu==1.9.0.post1
python-dotenv==1.0.1
